requirements.txt
pixoo == 0.9.2
Pillow == 10.4.0
